Rename ErrorPresenter --> ErrorMessage::Presenter
More accurately reflects its intention and groups
it with its dependendants.
