+++ app/form_builders/adp_text_field.rb
@@ -14,7 +14,7 @@ class AdpTextField
   #   * hint_text: Hint text displayed underneath the label
   #   * input_classes: Css classes on the input
   #   * input_type: Input type will default to `text`
-  #   * errors: An ErrorPresenter for the form object, or the form object itself
+  #   * errors: An ErrorMessage::Presenter for the form object, or the form object itself
   #   * error_key: supply to override the id used for both a page anchor tag and the error message translation lookup key
   #   * value: the value to display (if not specified, the value is taken by calling method on the form object)
   #   * id: the id to use for the field (if not specified, it is generated by snake casing the form object_name)
@@ -39,7 +39,7 @@ class AdpTextField
     @errors
   end
 
-  # @errors can either be an instance of ActiveModel::Errors or ErrorPresenter
+  # @errors can either be an instance of ActiveModel::Errors or ErrorMessage::Presenter
   def has_errors?
     return false if @errors.nil?
